Move blue.generic and origin tests from test_basic to test_blue_generic
